refactor: extract common pieces from unit list item components
