+++ components/ArmyRoster/ArmyRoster.tsx
@@ -2,7 +2,7 @@ import { type Army } from 'appdeptus/models'
 import React, { type ComponentProps, memo } from 'react'
 import { FlatList } from 'react-native'
 import { VStack } from '../ui'
-import UnitListItem from '../UnitListItem'
+import { UnitListItem } from '../UnitListItem'
 
 type ArmyRosterProps = {
   roster: Army['roster']
