+++ components/GameArmyRoster/GameArmyRoster.tsx
@@ -1,8 +1,8 @@
 import { type GameArmy } from 'appdeptus/models'
 import React, { type ComponentProps, memo } from 'react'
 import { FlatList } from 'react-native'
-import GameUnitListItem from '../GameUnitListItem'
 import { Pressable, VStack } from '../ui'
+import { GameUnitListItem } from '../UnitListItem'
 
 type GameArmyRosterProps = {
   roster: GameArmy['roster']
