+++ components/UnitListItem/UnitListItem.tsx
@@ -9,20 +9,19 @@ import {
   Car,
   CircleFadingPlus,
   Crown,
-  Link,
   type LucideIcon,
   Shield,
   UserRound,
   UsersRound
 } from 'lucide-react-native'
 import pluralize from 'pluralize'
-import { type ComponentProps, memo, type PropsWithChildren } from 'react'
+import { type ComponentProps, memo } from 'react'
 import Card from '../Card'
 import IconBadge from '../IconBadge'
-import InnerBorder from '../InnerBorder'
-import InsetShadow from '../InsetShadow'
 import Text from '../Text'
-import { HStack, Icon, VStack } from '../ui'
+import { HStack, VStack } from '../ui'
+import EmbarkedUnit from './EmbarkedUnit'
+import TeamUnit from './TeamUnit'
 
 type UnitListItemProps = {
   item: Army['roster'][number]
@@ -133,23 +132,10 @@ type TeamDetailProps = {
 }
 
 const TeamDetail = ({ team }: TeamDetailProps) => (
-  <VStack space='sm'>
-    <UnitDetail unit={team.leader} />
-    <Icon
-      as={Link}
-      className='px-4 color-primary-50'
-      size='md'
-    />
-    <UnitDetail unit={team.bodyguard} />
-  </VStack>
-)
-
-const EmbarkedUnit = ({ children }: PropsWithChildren) => (
-  <InnerBorder rounded='2xl'>
-    <InsetShadow>
-      <VStack className='rounded-2xl bg-primary-800 p-4'>{children}</VStack>
-    </InsetShadow>
-  </InnerBorder>
+  <TeamUnit
+    BodyGuard={<UnitDetail unit={team.bodyguard} />}
+    Leader={<UnitDetail unit={team.leader} />}
+  />
 )
 
 const unitTypeToIcon = {
