+++ components/UnitListItem/index.ts
@@ -1,1 +1,2 @@
-export { default } from './UnitListItem'
+export { default as GameUnitListItem } from './GameUnitListItem'
+export { default as UnitListItem } from './UnitListItem'
