+++ components/index.ts
@@ -16,7 +16,6 @@ export { default as FilterTopBar } from './FilterTopBar'
 export { default as GameArmyRoster } from './GameArmyRoster'
 export { default as GameBackground } from './GameBackground'
 export { default as GameDataTable } from './GameDataTable'
-export { default as GameUnitListItem } from './GameUnitListItem'
 export { default as IconBadge } from './IconBadge'
 export { default as Input } from './Input'
 export { default as InsetShadow } from './InsetShadow'
@@ -32,7 +31,7 @@ export { default as TabMenu } from './TabMenu'
 export { default as Text } from './Text'
 export { default as TextLink } from './TextLink'
 export { default as useToast } from './Toast'
-export { default as UnitListItem } from './UnitListItem'
+export * from './UnitListItem'
 export { default as UnitName } from './UnitName'
 export { default as VersusBackground } from './VersusBackground'
